refactor(pr-followup): the watch is a visible persistent monitor, gone only at terminal
