+++ plugin/skills/muggle-pr-followup/arm-watcher.md
@@ -1,11 +1,11 @@
 # Arming the Watch
 
-How an orchestrating session starts — or resumes — the watch on one PR. Every arming point runs this same sequence: [`bootstrap.md`](bootstrap.md) Step 8, [`auto-track.md`](auto-track.md) Step 6, and the executor's post-cycle respawn.
+How an orchestrating session starts — or restarts — the watch on one PR. Every arming point runs this same sequence: [`bootstrap.md`](bootstrap.md) Step 8, [`auto-track.md`](auto-track.md) Step 6, and the executor's post-cycle respawn.
 
 1. **Drain.** Run one tick per [`contract.md`](contract.md). It acts on everything already outstanding — actionable threads, body-only reviews past the watermark, a stale branch, red CI — and finalizes a terminal PR. If the tick dispatched a cycle, stop here: the cycle's exit path re-arms when it finishes.
-2. **Watch.** PR still open and nothing dispatched → on the first arm, spawn the [`pr-watcher`](../../agents/pr-watcher.md) agent named `PR #<n> — <title>`; on every later arm, send that same agent a resume message. One watcher per PR, for the PR's whole life: it baselines when it wakes, reports on new comment activity / a terminal PR / a failed fetch, then pauses until resumed. If the agent no longer exists — its session ended, or it was killed — spawn a fresh one: the poller holds nothing durable, so everything it needs is re-derived from the slot and live provider state.
-3. **On report.** `NEW_COMMENT` or `TERMINAL` → step 1: the tick derives what changed from live provider state, and a terminal PR finalizes there. `ERROR` → surface the reason; [`reconcile.md`](reconcile.md) re-arms the slot on its next sweep.
+2. **Watch.** PR still open and nothing dispatched → start the watch as a **persistent background monitor** in the orchestrating session, labeled `PR #<n> — <title>`. The monitor is the watch's visible handle: it shows as a running task for as long as the PR is being polled, and it ends — and disappears — only when it reports activity, the PR goes terminal, or fetches persistently fail. Its loop checks about every 60 seconds, against the watermark captured right after the drain — pass the watermark in; a loop that captures its own baseline swallows anything that landed between drain and launch. On a newer submitted review, a newer thread comment, a thread newly unresolved, a terminal PR, or five consecutive failed fetches, it prints one line saying which and exits in the same breath. Quiet iterations print nothing and cost nothing — no model tokens are spent while the watch is quiet.
+3. **On event.** The monitor's exit line wakes the session: activity or a terminal PR → step 1, where the tick derives everything from live provider state and a terminal PR finalizes. Persistent fetch failure → surface the reason; [`reconcile.md`](reconcile.md) re-arms the slot on its next sweep.
 
-Every cycle ends with this arming — including cycles the orchestrator started without a watcher report. A cycle that skips the resume leaves the watcher's baseline stale, and its next detection will be the loop's own reply.
+Every cycle ends with this arming — including cycles the orchestrator started without a watch event. A cycle that skips the re-arm leaves the watermark stale, and the next detection will be the loop's own reply.
 
-Drain-then-watch is the invariant, and it is why the drain precedes every resume, not just the first: the watcher re-baselines when it wakes, so feedback that landed while a cycle ran would slip under the fresh baseline. The tick is the one component that checks live state with the real semantics — marker rule, watermark, attempt budgets — that the poller deliberately lacks.
+Drain-then-watch is the invariant, and it is why the drain precedes every arm, not just the first: the monitor only needs to see what arrives after its watermark. The tick is the one component that checks live state with the real semantics — marker rule, watermark, attempt budgets — that the watch deliberately lacks.
